+++ templates/super_minimal.html
@@ -1,0 +1,8 @@
+<!DOCTYPE html>
+<html>
+<head><title>Super Minimal</title></head>
+<body>
+    <h1>Hello from Template!</h1>
+    <p>This is a test.</p>
+</body>
+</html>
